+++ src/app/Http/Resources/GroupUserInvitationResource.php
@@ -1,0 +1,22 @@
+<?php declare(strict_types=1);
+
+namespace App\Http\Resources;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class GroupUserInvitationResource extends JsonResource
+{
+    /**
+     * @param  \Illuminate\Http\Request  $request
+     * @return array
+     */
+    public function toArray($request)
+    {
+        return [
+            'id' => $this->id,
+            'email' => $this->email,
+            'invitation_code' => $this->invitation_code,
+            'status' => $this->status,
+        ];
+    }
+}