+++ src/resources/js/pages/sessions/register/info.vue
@@ -1,5 +1,5 @@
 <template>
-    <layout :sut="session.sut" :components="components">
+    <layout :components="components" :session="session">
         <form @submit.prevent="submit" class="col-12">
             <div class="card">
                 <div class="row">
@@ -52,12 +52,19 @@
                     </div>
                     <div class="col-6">
                         <div class="card-header pl-0 border-0">
-                            <h3 class="card-title">Select use cases</h3>
+                            <h3 class="card-title">
+                                {{
+                                    isCompliance
+                                        ? 'Selected use cases'
+                                        : 'Select use cases'
+                                }}
+                            </h3>
                         </div>
                         <div class="card-body pt-0 pl-0">
                             <test-case-checkboxes
-                                style="max-height: 320px;"
+                                style="max-height: 320px"
                                 :useCases="useCases"
+                                :isCompliance="isCompliance"
                                 v-model="form.test_cases"
                             />
                             <div
@@ -115,6 +122,7 @@ export default {
     data() {
         return {
             sending: false,
+            isCompliance: this.session.type === 'compliance',
             form: {
                 name: this.session.info ? this.session.info.name : null,
                 description: this.session.info
@@ -125,6 +133,17 @@ export default {
                     : [],
             },
         };
+    },
+    created() {
+        let form = this.form;
+
+        if (this.isCompliance && !this.session.info) {
+            this.useCases.data.forEach(function (useCase) {
+                useCase.testCases.forEach(function (testCase) {
+                    form.test_cases.push(testCase.id);
+                });
+            });
+        }
     },
     methods: {
         submit() {