+++ src/routes/web.php
@@ -46,6 +46,10 @@ Route::namespace('Groups')->group(function () {
         'groups.environments',
         'GroupEnvironmentController'
     )->except(['show']);
+    Route::resource(
+        'groups.user-invitations',
+        'GroupUserInvitationController'
+    )->except(['show', 'edit']);
 });
 
 /**
@@ -94,12 +98,31 @@ Route::name('sessions.')
         Route::name('register.')
             ->prefix('register')
             ->group(function () {
+                Route::get('type', 'RegisterController@showTypeForm')->name(
+                    'type'
+                );
+                Route::post(
+                    'type/{type}',
+                    'RegisterController@storeType'
+                )->name('type.store');
                 Route::get('sut', 'RegisterController@showSutForm')->name(
                     'sut'
                 );
                 Route::post('sut', 'RegisterController@storeSut')->name(
                     'sut.store'
                 );
+                Route::get(
+                    'questionnaire/summary',
+                    'RegisterController@questionnaireSummary'
+                )->name('questionnaire.summary');
+                Route::get(
+                    'questionnaire/{section}',
+                    'RegisterController@showQuestionnaireForm'
+                )->name('questionnaire');
+                Route::post(
+                    'questionaire/{section}',
+                    'RegisterController@storeQuestionnaire'
+                )->name('questionnaire.store');
                 Route::get('info', 'RegisterController@showInfoForm')->name(
                     'info'
                 );
@@ -240,6 +263,17 @@ Route::name('admin.')
                     'group-candidates',
                     'TestCaseController@groupCandidates'
                 )->name('group-candidates');
+            });
+        Route::name('questionnaire.')
+            ->prefix('questionnaire')
+            ->group(function () {
+                Route::get(
+                    'import',
+                    'QuestionnaireController@showImportForm'
+                )->name('import');
+                Route::post('import', 'QuestionnaireController@import')->name(
+                    'import.confirm'
+                );
             });
     });